Code adjustments to work in parallel.
A major issue with the previous version was that adequate memory for FFTW MPI
routines was not available. This was causing segfaults, since the FFTW routines
were expecting the memory to be there. This has been fixed, and a lot of
"global" variables have been replaced with their corresponding "process"
variables to make the MPI implementation work. This version has been checked
for accuracy with the Taylor-Green vortex problem on 1 and 4 processes.